requirements.txt
fastapi
uvicorn
httpx
feedparser
pydantic
pydantic-settings
python-dotenv
openai
google-genai
qdrant-client